Implement peer verification, authentication
New verification system contains a wizard that will show
on one peer and prompt for security code on the other.

The security code is showed in the wizard and the user must enter
it on the other peer. If the code matches, the two peers are paired,
remembering each other's identity.

Certificate trust is no longer checked. The only thing that matters
is that peer has certificate that is in the local pool of authorized
nodes.

Conversations may now select required level of authentication, so that
introduction does not need any and clipboard sync does.

ConnectionManager exposes signals and slots for controlling
the peer verification in UIs.