Node ids, Settings signals for node add/update, addOrUpdateNode method
Due to a need to identify nodes (update and other operations), every
node has assigned a unique id. These ids are generated on runtime,
nothing is saved to config.

When a node is added to the Settings, it fires either signal nodeUpdated,
if the node is already in pool (e.g. during re-verification)
or nodeAdded, if it isn't there yet. This is handled by method
addOrUpdateNode(), which replaces addNode().