version> Fix: "libhaveclipcore.a" is always re‑created at build time
Before this commit:

    GITVERSION = src/git_version.h

Current Working Dir for `make` invocation is $${OUT_PWD} (for example, see the `Sender.o` target in the generated Makefile).
So, in the target, "src/git_version.h" is equal to "$$OUT_PWD/src/git_version.h", but real target is "$$PWD/src/git_version.h":

    versiontarget.target = $$GITVERSION
    versiontarget.commands = $$PWD/utils/git_version.sh \"$$PWD/../\" \"$$PWD/$$GITVERSION\"
    versiontarget.depends = FORCE
    PRE_TARGETDEPS += $$GITVERSION
    QMAKE_EXTRA_TARGETS += versiontarget

Results (after the commit): `git_version.sh` is always invoked, and "libhaveclipcore.a" is re-created if "git_version.h" (included in "qmlclipboardmanager.cpp") is updated.

Also in the Pull Request:
- Preventing unnecessary `git_version.sh` invocations.
- Alternative to `git_version.sh`.