version> Rename files…
… before the main commit: Set `VERSION` in one place to prevent version mismatch.

  git mv src/Version.h src/Version.pri.h

It was extracted into the separate commit because the changes of the main commit are too large for the file rename detection heuristic to work (with the default value: `-M50%` - `git diff/log -M/--find-renames`; `-C20` - `git blame -C`). And the correct detection of file renaming is necessary to continue (preserve) its history.